Refactored package docs generation logic
