--- src/main-handlers/package/build-docs.js
@@ -1,6 +1,0 @@
-module.exports = ({ replaceInFile, fs }, templatePath, main) => {
-  if (!fs.existsSync(templatePath) || !templatePath) {
-    return main
-  }
-  return replaceInFile(fs.readFileSync(templatePath, 'utf8'), 'main', main)
-}
