+++ src/main-handlers/package/index.js
@@ -1,14 +1,19 @@
 const modules = {
   fs: require('fs'),
-  path: require('path'),
-  replaceInFile: require('../../helpers/replace-in-file.js')
+  replaceInFile: require('../../helpers/replace-in-file.js'),
+  glob: require('glob'),
+  jsdoc2md: require('jsdoc-to-markdown')
 }
-const getMain = require('./get-main.js')
-const buildDocs = require('./build-docs.js')
+
+const path = require('path')
 
 module.exports = templatePath => {
   console.log('INFO: generating documentation from provided template and all JS files in repository...')
-  const docs = buildDocs(modules, templatePath, getMain(modules))
+  const data = {
+    ...require('./get-data.js')(modules),
+    ...require('../../helpers/jsdoc-utils/get-data.js')(modules)
+  }
+  const docs = require('../../helpers/jsdoc-utils/build-docs.js')(modules, data, path.join(__dirname, 'package-partial.md'), templatePath)
   console.log('SUCCESS: documentation successfully generated!')
   modules.fs.writeFileSync('README.md', docs, 'utf8')
 }
